fix(renderer): degrade instead of dying on non-triangulable VBAP geometry
A per-speaker frequency crossover can leave a band with a speaker set the
VBAP panner can't triangulate (e.g. a sub-bass band of LFE + two surrounds,
with the LFE at the listener). find_ls_triplets then failed, the error
propagated up, and orender_create returned NULL — so an mpv host dropped the
track and played NO audio, with only a stderr line. Silent, total failure for
a config that's merely degenerate.

Two changes:

1. Never fail silently. At both VBAP-panner build sites (the generic
   backend_registry plan and the spatial_renderer construction path — covering
   the main layout, every crossover band, and hybrid inner backends), a
   triangulation failure now falls back to the degenerate directional backend
   and logs a loud warn! naming the speakers and reason. The warning surfaces
   on stderr AND in Studio's log panel (liborender routes log::* over OSC), and
   audio keeps playing. Build-time only — no per-sample cost.

2. Extend the degenerate backend to N speakers with a real pan. Rename
   FewSpeakerBackend → DegenerateVbapBackend and generalise it from "1-2
   speakers, equal-power above" to any N: pan by local pairwise VBAP between the
   two nearest speaker directions (well-defined for collinear/coplanar sets the
   full panner rejects), keeping the 1- and 2-speaker behaviour. Speakers at the
   listener (no usable direction, e.g. a spatialised LFE) are flagged omni and
   take a constant baseline share so they still get signal.

Reproduced the original find_ls_triplets abort with the user's 18-speaker
crossover config; with the fix the [0,40)Hz band logs the degraded-fallback
warning and the engine builds through every band. Adds collinear + omni unit
tests; existing crossover/unified-table tests still pass.
